requirements.txt
# Core runtime
typer==0.12.5
python-dotenv==1.0.1

# LLM clients
openai>=1.30.0
requests>=2.31.0
httpx>=0.27.0          

# PDF report generator (tools/report_pdf.py)
reportlab>=4.1.0
matplotlib>=3.8.0
numpy>=1.26.0
